HW-4 Добавил новые методы и заменил старые методы на Stream, поменял структуру в проекте
