+++ .idea/.gitignore
@@ -1,3 +1,7 @@
 # Default ignored files
 /shelf/
 /workspace.xml
+/modules/
+/compiler.xml
+gradle.xml
+
