+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/homework.main.iml" filepath="$PROJECT_DIR$/.idea/modules/homework.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/mts-homework-dcherednikov.iml" filepath="$PROJECT_DIR$/.idea/mts-homework-dcherednikov.iml" />
     </modules>
   </component>
