+++ homework/settings.gradle.kts
@@ -1,0 +1,3 @@
+rootProject.name = "homework"
+include("src:main:ru.mts.homeworks")
+findProject(":src:main:ru.mts.homeworks")?.name = "ru.mts.homeworks"
