+++ homework/src/main/java/ru/mts/homeworks/exception/EmptyAnimalMapException.java
@@ -1,0 +1,9 @@
+package ru.mts.homeworks.exception;
+
+public class EmptyAnimalMapException extends RuntimeException{
+    private static final String ERR_MESSAGE = "На вход пришел пустой список животных!";
+
+    public EmptyAnimalMapException(){
+        super(ERR_MESSAGE);
+    }
+}
