+++ homework/src/main/java/ru/mts/homeworks/exception/InvalidAnimalBirthDateException.java
@@ -1,0 +1,10 @@
+package ru.mts.homeworks.exception;
+
+public class InvalidAnimalBirthDateException extends Exception{
+    public InvalidAnimalBirthDateException(String message){
+        super(message);
+    }
+    public InvalidAnimalBirthDateException(String message, Throwable cause){
+        super(message,cause);
+    }
+}
