+++ homework/src/main/java/ru/mts/homeworks/exception/InvalidAnimalException.java
@@ -1,0 +1,7 @@
+package ru.mts.homeworks.exception;
+
+public class InvalidAnimalException extends RuntimeException{
+    public InvalidAnimalException(String message){
+        super(message);
+    }
+}
