+++ homework/src/main/java/ru/mts/homeworks/service/Animal.java
@@ -1,0 +1,14 @@
+package ru.mts.homeworks.service;
+
+
+import java.time.LocalDate;
+
+public interface Animal{
+    String getBreed();
+    String getName();
+    Double getCost();
+    String getCharacter();
+    LocalDate getBirthDate();
+    String getClassName();
+    int getAge();
+}
