+++ homework/src/main/java/ru/mts/homeworks/service/SearchService.java
@@ -1,0 +1,8 @@
+package ru.mts.homeworks.service;
+
+import ru.mts.homeworks.model.*;
+import ru.mts.homeworks.exception.*;
+
+public interface SearchService {
+    void checkLeapYearAnimal(AbstractAnimal animal) throws InvalidAnimalBirthDateException;
+}
