--- lesson1/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Tue Feb 20 11:31:07 MSK 2024
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.0-bin.zip
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
