--- lesson1/src/HW_1/Animal.java
@@ -1,11 +1,0 @@
-package HW_1;
-
-import java.time.LocalDate;
-
-public interface Animal{
-    String getBreed();
-    String getName();
-    Double getCost();
-    String getCharacter();
-    LocalDate getBirthDate();
-}
