--- lesson1/src/HW_1/Dog.java
@@ -1,9 +1,0 @@
-package HW_1;
-
-import java.time.LocalDate;
-
-public class Dog extends Pet{
-    public Dog(String name, LocalDate birthDate,String breed, double cost, String character){
-        super(name, birthDate, breed, cost, character);
-    }
-}
