--- lesson1/src/HW_1/EmptyAnimalMapException.java
@@ -1,9 +1,0 @@
-package HW_1;
-
-public class EmptyAnimalMapException extends RuntimeException{
-    private static final String ERR_MESSAGE = "На вход пришел пустой список животных!";
-
-    public EmptyAnimalMapException(){
-        super(ERR_MESSAGE);
-    }
-}
