--- lesson1/src/HW_1/InvalidAnimalBirthDateException.java
@@ -1,10 +1,0 @@
-package HW_1;
-
-public class InvalidAnimalBirthDateException extends Exception{
-    public InvalidAnimalBirthDateException(String message){
-        super(message);
-    }
-    public InvalidAnimalBirthDateException(String message, Throwable cause){
-        super(message,cause);
-    }
-}
