--- lesson1/src/HW_1/InvalidAnimalException.java
@@ -1,7 +1,0 @@
-package HW_1;
-
-public class InvalidAnimalException extends RuntimeException{
-    public InvalidAnimalException(String message){
-        super(message);
-    }
-}
