--- lesson1/src/HW_1/SearchService.java
@@ -1,5 +1,0 @@
-package HW_1;
-
-public interface SearchService {
-    void checkLeapYearAnimal(AbstractAnimal animal) throws InvalidAnimalBirthDateException;
-}
