--- lesson1/src/HW_1/Shark.java
@@ -1,9 +1,0 @@
-package HW_1;
-
-import java.time.LocalDate;
-
-public class Shark extends Predator{
-    public Shark(String name, LocalDate birthDate, String breed, String character, String area) {
-        super(name, birthDate,breed, character,area);
-    }
-}
