--- lesson1/src/HW_1/Wolf.java
@@ -1,9 +1,0 @@
-package HW_1;
-
-import java.time.LocalDate;
-
-public class Wolf extends Predator{
-    public Wolf(String name, LocalDate birthDate, String breed, String character, String area) {
-        super(name,birthDate ,breed, character, area);
-    }
-}
